pyproject.toml
[tool.poetry]
name = "reair"
version = "0.1.0"
description = ""
authors = ["Chapalax <max250903@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
